requirements.txt
backports-abc==0.5
backports.shutil-get-terminal-size==1.0.0
backports.weakref==1.0rc1
bleach==1.5.0
certifi==2017.1.23
click==6.7
configparser==3.5.0
cycler==0.10.0
decorator==4.0.11
entrypoints==0.2.2
enum34==1.1.6
funcsigs==1.0.2
functools32==3.2.3.post2
html5lib==0.9999999
intervaltree==2.1.0
ipdb==0.10.3
ipykernel==4.6.1
ipython==5.3.0
ipython-genutils==0.2.0
ipywidgets==6.0.0
jedi==0.10.2
Jinja2==2.9.6
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.0.1
jupyter-console==5.1.0
jupyter-core==4.3.0
magenta==0.1.10
Markdown==2.6.8
MarkupSafe==1.0
matplotlib==2.0.0
mido==1.2.4
mistune==0.7.4
mock==2.0.0
nbconvert==5.1.1
nbformat==4.3.0
notebook==5.0.0
numpy==1.13.1
olefile==0.44
pandocfilters==1.4.1
pathlib2==2.2.1
pbr==3.1.1
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.1.0
pretty-midi==0.2.7
prompt-toolkit==1.0.14
protobuf==3.3.0
ptyprocess==0.5.1
pyaml==16.12.2
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.6.0
python-rtmidi==1.0.0
pytz==2017.2
PyYAML==3.12
pyzmq==16.0.2
qtconsole==4.3.0
scandir==1.5
scipy==0.19.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
sortedcontainers==1.5.7
subprocess32==3.2.7
tensorflow==1.2.1
terminado==0.6
testpath==0.3
tornado==4.4.3
traitlets==4.3.2
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.12.2
widgetsnbextension==2.0.0
